pyproject.toml
[project]
name = "lol-commentary"
version = "0.1.0"
description = "AI-powered League of Legends commentary system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp>=2024.0.0",
    "opencv-python-headless>=4.9.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "youtube-transcript-api>=0.6.0",
    "riotwatcher>=3.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lol_commentary"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
